change package structure files locations
