+++ bower.json
@@ -6,11 +6,22 @@
     "Allen Royston <allen.acr@gmail.com>",
     "Orif N. Jr. <orif.zade@gmail.com>"
   ],
-  "description": "An Angular module for the awesome Croppie.js",
+  "description": "An awesome image cropping and rotating module for AngularJS",
   "main": [
-    "unminified/ng-croppie.js",
-    "unminified/ng-croppie.css"
+    "ng-croppie.js",
+    "ng-croppie.css"
   ],
+  "dependencies": {
+    "croppie": "2.4.1"
+  },
+  "overrides": {
+    "croppie": {
+      "main": [
+        "croppie.js",
+        "croppie.css"
+      ]
+    }
+  },
   "moduleType": [],
   "keywords": [
     "angular",
