+++ ng-croppie.css
@@ -75,15 +75,15 @@
 /*************************************/
 /***** STYLING RANGE INPUT ***********/
 /*************************************/
-/* http://brennaobrien.com/blog/2014/05/style-input-type-range-in-every-browser.html */
+/*http://brennaobrien.com/blog/2014/05/style-input-type-range-in-every-browser.html */
 /*************************************/
 
 .cr-slider {
     -webkit-appearance: none;
-    /*removes default webkit styles*/
-    /*border: 1px solid white; *//*fix for FF unable to apply focus style bug */
+/*removes default webkit styles*/
+	/*border: 1px solid white; *//*fix for FF unable to apply focus style bug */
     width: 300px;
-    /*required for proper track sizing in FF*/
+/*required for proper track sizing in FF*/
     max-width: 100%;
 }
 
@@ -110,7 +110,7 @@
 }
 /*
 .cr-slider:focus::-webkit-slider-runnable-track {
-    background: #ccc;
+background: #ccc;
 }
 */
 
@@ -131,7 +131,7 @@
     margin-top: -6px;
 }
 
-/* hide the outline behind the border */
+/*hide the outline behind the border*/
 .cr-slider:-moz-focusring {
     outline: 1px solid white;
     outline-offset: -1px;
@@ -140,32 +140,33 @@
 .cr-slider::-ms-track {
     width: 100%;
     height: 5px;
-    background: transparent; /* remove bg colour from the track, we'll use ms-fill-lower and ms-fill-upper instead */
-    border-color: transparent; /* leave room for the larger thumb to overflow with a transparent border */
-    border-width: 6px 0;
-    color: transparent;/* remove default tick marks */
+    background: transparent;
+/*remove bg colour from the track, we'll use ms-fill-lower and ms-fill-upper instead */
+	border-color: transparent;/*leave room for the larger thumb to overflow with a transparent border */
+	border-width: 6px 0;
+	color: transparent;/*remove default tick marks*/
 }
 .cr-slider::-ms-fill-lower {
-    background: rgba(0, 0, 0, 0.5);
-    border-radius: 10px;
+	background: rgba(0, 0, 0, 0.5);
+	border-radius: 10px;
 }
 .cr-slider::-ms-fill-upper {
-    background: rgba(0, 0, 0, 0.5);
-    border-radius: 10px;
+	background: rgba(0, 0, 0, 0.5);
+	border-radius: 10px;
 }
 .cr-slider::-ms-thumb {
-    border: none;
-    height: 16px;
-    width: 16px;
-    border-radius: 50%;
-    background: #ddd;
-    margin-top:1px;
+	border: none;
+	height: 16px;
+	width: 16px;
+	border-radius: 50%;
+	background: #ddd;
+	margin-top:1px;
 }
 .cr-slider:focus::-ms-fill-lower {
-    background: rgba(0, 0, 0, 0.5);
+	background: rgba(0, 0, 0, 0.5);
 }
 .cr-slider:focus::-ms-fill-upper {
-    background: rgba(0, 0, 0, 0.5);
+	background: rgba(0, 0, 0, 0.5);
 }
 /*******************************************/
 
@@ -173,24 +174,24 @@
 /* Rotation Tools */
 /***********************************/
 .cr-rotate-controls {
-    position: absolute;
-    bottom: 5px;
-    left: 5px;
-    z-index: 1;
+	position: absolute;
+	bottom: 5px;
+	left: 5px;
+	z-index: 1;
 }
 .cr-rotate-controls button {
-    border: 0;
-    background: none;
+	border: 0;
+	background: none;
 }
 .cr-rotate-controls i:before {
-    display: inline-block;
-    font-style: normal;
-    font-weight: 900;
-    font-size: 22px;
+	display: inline-block;
+	font-style: normal;
+	font-weight: 900;
+	font-size: 22px;
 }
 .cr-rotate-l i:before {
-    content: '↺';
+	content: '↺';
 }
 .cr-rotate-r i:before {
-    content: '↻';
+	content: '↻';
 }
